requirements.txt
streamlit
tensorflow
keras
numpy
Pillow
matplotlib